change rule structure so synthsr only clones once
