+++ autoafids/workflow/Snakefile
@@ -112,39 +112,8 @@ if stereotaxy_enabled:
 
 
 if config["modality"] != "T1w":
-
-    rule synthsr:
-        input:
-            im=bids(
-                root=str(Path(config["bids_dir"])),
-                datatype="anat",
-                suffix=f"{config['modality']}.nii.gz",
-                **inputs[config["modality"]].wildcards,
-            ),
-        output:
-            SynthSR=bids(
-                root=work,
-                datatype="SynthSR",
-                desc=f"{config['modality']}-SynthSR",
-                suffix="T1w.nii.gz",
-                **inputs[config["modality"]].wildcards,
-            ),
-        log:
-            bids(
-                root="logs",
-                suffix="synthsr_output.txt",
-                **inputs[config["modality"]].wildcards
-            ),
-        resources:
-            mem_mb=10000,
-        threads: max(1, workflow.cores // max(1, available_mem_mb // 10000))
-        params:
-            modality=config["modality"],
-            download_dir=download_dir,
-        conda:
-            "envs/synthsr.yaml"
-        script:
-            "scripts/run_synthsr.py"
+    
+    include: "rules/synthsr.smk"
 
 
 if "n4_bias_correction" in profile_settings and config["modality"] == "T1w":
